Wongle Bucket + Textures
Weighted bucket + various texture maps including a height map.

Due to issues on my end I have not been able to assign these textures to a material. Something is playing up and I am not sure why at this stage, so a new material will need to be made for the bucket manually containing the included maps.