Wongle Axe Additions + Animations
Contains two axes; one for the purpose of cutting wood, the other is a direct weapon upgrade based on the infection and is not designed for cutting wood.

Included are two weighted models, various texture maps (including emissives that will need to be applied separately) and a wood cutting animation for the Wongle Worker.