Added summative folder + latest summative for milestones
Tomorrow as a group we should look at this in the morning, discuss all features and bugs, and format them appropriately in a google doc for submission on Friday.

I (Ben L) Will be off to another class in the afternoon and away from Friday so let's tackle this early.