Alternate Roughness Map
I've heard that the weird oily skin issue might be caused in by a conflict with how Unity handles roughness. While the maps appear fine in Maya and Substance, Unity seems to confuse itself when handling roughness maps. Apparently this is because Unity reads this information in inverse. To test this I have inverted a single roughness map to see if it resolves the issue.