Bundy fixes for spawning on new map
Spherical/circular styled map bugs out the bundy respawner. Had to create a shit plane to minic th region then put a class on the objs spawning which then raycast down to check whether the ground is there. Could use a distance from house, if house is at centerpoint and check within the radius of map.