Idle Select Camera Zoom Bug
Camera will now un-zoom provided that the camera's Y val isn't set to the default of '50'. So when selecting idle workers they will always be in your view. opposed to before with random zooms, could have used the lookAt function, but becuause we aren't yet lerping the cam rotation this would break functionality....