Weird AudioHandler bug + more
Sounds won't play once the clip has been updated, not to sure why.. I can see that the source is being altered (clip is changing) but no sound. Altered from resources.load to a standard inspector/editor load, and still encountered the same problem.
Revised AudioHandler to be less recursive - it used to have multiple definitions and each would load all sounds. Now it just loads all sounds once, and a new function takes the audiosource to edit (wongleworker) and updates it -> needs more testing, but I believe this is the way to go.