Implemented Time Scale Adjustment Cheat + Warning notif
Added a function to Charmaines notifiction system, couldnt be bothered updated the 23 or so functions that reference the original function. Basically just adds a "WARNING: + message" with red text. Added function to adjust fadetime for when player is in speed adjustment cheat

Added speed adjustment cheat which speeds the games timescale up 3x. cheat can be accessed by typing "boosttimepls"