Optimized Construction Script + Prefabs
Was attempting to add some new prefabs and noticed some optimizations could be made to the construction script.

Prefabs are being buggy af, I got Ben L to alter the pivot points of the obj so they wouldnt place in the ground, which they dont . But when being placed into a separate prefab as a child their pivot points become cooked. cbf altering new prefabs pivot point in blender and re-exporting...