Wongle Farm Builder bug fix + WongleController cleanup
Fixed a bug where the wongle which would go to construct the farm (instantly constructs) gets stuck in construction mode -> Could create functionality to make the wongle farm as soon as it's placed...

Cleaned up the WongleController Update function. Split it up into functions.

Found a bug with the new geiger crystal, the main prefab is dissapearing -> will sort tomorrow, replaced with the old one for build...